+++ chromium/components/autofill/core/browser/payments/strike_database_integrator_base.h
@@ -5,6 +5,8 @@
 #ifndef COMPONENTS_AUTOFILL_CORE_BROWSER_PAYMENTS_STRIKE_DATABASE_INTEGRATOR_BASE_H_
 #define COMPONENTS_AUTOFILL_CORE_BROWSER_PAYMENTS_STRIKE_DATABASE_INTEGRATOR_BASE_H_
 
+#include <stdint.h>
+
 #include "components/autofill/core/browser/payments/strike_database.h"
 
 namespace autofill {
@@ -55,6 +57,8 @@ class StrikeDatabaseIntegratorBase {
  protected:
   // Removes all strikes in which it has been longer than GetExpiryTimeMicros()
   // past |last_update_timestamp|.
+  // TODO(crbug/1061639): Provide option to NOT expire strikes, perhaps by
+  //                      GetExpiryTimeMicros() return a base::Optional.
   void RemoveExpiredStrikes();
 
  private:
@@ -95,7 +99,7 @@ class StrikeDatabaseIntegratorBase {
   virtual int GetMaxStrikesLimit() = 0;
 
   // Returns the time after which the most recent strike should expire.
-  virtual long long GetExpiryTimeMicros() = 0;
+  virtual int64_t GetExpiryTimeMicros() = 0;
 
   // Returns whether or not a unique string identifier is required for every
   // strike in this project.
